+++ src/test/java/org/jabref/logic/crawler/StudyYamlParserTest.java
@@ -2,7 +2,6 @@ package org.jabref.logic.crawler;
 
 import java.net.URL;
 import java.nio.file.Path;
-import java.time.LocalDate;
 import java.util.List;
 
 import org.jabref.logic.util.io.FileUtil;
@@ -35,22 +34,18 @@ class StudyYamlParserTest {
                 new StudyDatabase("Medline/PubMed", true), new StudyDatabase("IEEEXplore", false));
 
         expectedStudy = new Study(authors, studyName, researchQuestions, queryEntries, libraryEntries);
-        expectedStudy.setLastSearchDate(LocalDate.parse("2020-11-26"));
     }
 
     @Test
     public void parseStudyFileSuccessfully() throws Exception {
         Study study = new StudyYamlParser().parseStudyYamlFile(testDirectory.resolve("study.yml"));
-
         assertEquals(expectedStudy, study);
     }
 
     @Test
     public void writeStudyFileSuccessfully() throws Exception {
         new StudyYamlParser().writeStudyYamlFile(expectedStudy, testDirectory.resolve("study.yml"));
-
         Study study = new StudyYamlParser().parseStudyYamlFile(testDirectory.resolve("study.yml"));
-
         assertEquals(expectedStudy, study);
     }
 }